Implemented a wathdog for the DHT22 sensor
The DHT22 sensor is not responding from time to time
The only thing that helps is unpowering the sensor
But this cannot be done by rebooting the raspi
So I connected the 3.3V power source of the DHT22 sensor
to a relais that is controlled by GPIO21 port.
0: power on
1: power off

Note it would by possible doing this by a transistor, or schmitttrigger
but you'l loose 0.7V and I'm not sure if the DTH22 would work
and in anyway I may wan't to controll a fan or something alse with the
second relay block.